refactoring selection view model
