+++ app/src/main/java/com/songlib/presentation/screens/selection/step1/Step1Content.kt
@@ -27,7 +27,6 @@ fun Step1Content(
     ) {
         LazyColumn(
             modifier = Modifier.fillMaxSize(),
-//            verticalArrangement = Arrangement.spacedBy(5.dp),
             contentPadding = PaddingValues(horizontal = 5.dp)
         ) {
             items(books) { book ->
